refactor `aliases` from typescript config
