+++ src/components/index/Hero.tsx
@@ -1,8 +1,8 @@
 import Image from 'next/image'
 import Link from 'next/link'
 
-import Header from '../ui/Header/Header'
-import Button from '../ui/Button'
+import Header from '#/Global/Header/Header'
+import Button from '#/UI/Button'
 
 import HeroBackground from '%/hero/background.webp'
 import HeroText from '%/hero/text.svg'
@@ -31,7 +31,7 @@ function HeroSocials() {
   )
 }
 
-const screenHeight = 'h-[100vh]'
+const screenHeight = 'h-[100vh] h-[100svh]'
 
 export default function Hero() {
   return (
