+++ src/components/index/SchemaSVG.tsx
@@ -1,3 +1,5 @@
+'use client'
+
 import {useState, useEffect, useRef} from 'react'
 
 interface SchemaSVGProps {
